Change Float to Double

+++ src/Flowbox/GuiMockup/ShortStraw.hs
@@ -22,14 +22,14 @@ import           Flowbox.GuiMockup.JSInterop
 import           Debug.Trace
 
 
-shortStraw :: V.Vector (V2 Float) -> V.Vector (V2 Float)
+shortStraw :: V.Vector (V2 Double) -> V.Vector (V2 Double)
 shortStraw points = V.backpermute resampled cornersIndices
     where
         s = determineResampleSpacing points
         resampled = resample s points
         cornersIndices = getCorners resampled
 
-determineResampleSpacing :: V.Vector (V2 Float) -> Float
+determineResampleSpacing :: V.Vector (V2 Double) -> Double
 determineResampleSpacing points = diagonal / 40
     where
         diagonal = distance topLeft bottomRight
@@ -41,7 +41,7 @@ determineResampleSpacing points = diagonal / 40
         maxX = V.maximumBy (compare `on` (^. _x)) points ^. _x
         maxY = V.maximumBy (compare `on` (^. _y)) points ^. _y
 
-getCorners :: V.Vector (V2 Float) -> V.Vector Int
+getCorners :: V.Vector (V2 Double) -> V.Vector Int
 getCorners points = runST $ do
     let w = 3
         pointsLength = V.length points
@@ -101,7 +101,7 @@ removeAt n l = case back of
 remove :: Eq a => a -> [a] -> [a]
 remove toDelete l = [ x | x <- l, x /= toDelete]
 
-postProcessCorners :: V.Vector (V2 Float) -> [Int] -> V.Vector Float -> [Int]
+postProcessCorners :: V.Vector (V2 Double) -> [Int] -> V.Vector Double -> [Int]
 postProcessCorners points corners straws = runST $ do
     continue <- newSTRef False
     cornersRef <- newSTRef corners
@@ -149,7 +149,7 @@ postProcessCorners points corners straws = runST $ do
 
     readSTRef cornersRef
 
-halfwayCorner :: V.Vector Float -> Int -> Int -> Int
+halfwayCorner :: V.Vector Double -> Int -> Int -> Int
 halfwayCorner straws a b = runST $ do
     minValue <- newSTRef $ 1 / 0
     minIndex <- newSTRef 0
@@ -169,14 +169,14 @@ increment ref = modifySTRef ref (+1)
 decrement :: Num a => STRef s a -> ST s ()
 decrement ref = modifySTRef ref (subtract 1)
 
-isLine :: V.Vector (V2 Float) -> Int -> Int -> Bool
+isLine :: V.Vector (V2 Double) -> Int -> Int -> Bool
 isLine points p1 p2 = dist / pathDist > threshold
     where
         threshold = 0.95
         dist = distance (points V.! p1) (points V.! p2)
         pathDist = pathDistance points p1 p2
 
-pathDistance :: V.Vector (V2 Float) -> Int -> Int -> Float
+pathDistance :: V.Vector (V2 Double) -> Int -> Int -> Double
 pathDistance points i1 i2 = runST $ do
     d <- newSTRef 0
 
@@ -200,10 +200,10 @@ median vec' = runST $ do
     else
         MV.read vec ix
 
-resample :: Float -> V.Vector (V2 Float) -> V.Vector (V2 Float)
+resample :: Double -> V.Vector (V2 Double) -> V.Vector (V2 Double)
 resample i v@(V.toList -> points) = V.fromList $ go 0 (head points) (tail points) [head points]
     where
-        go :: Float -> V2 Float -> [V2 Float] -> [V2 Float] -> [V2 Float]
+        go :: Double -> V2 Double -> [V2 Double] -> [V2 Double] -> [V2 Double]
         go bigD previousPoint (point:ps) newPoints =
             let d = distance previousPoint point
             in  if bigD + d >= i
@@ -212,10 +212,10 @@ resample i v@(V.toList -> points) = V.fromList $ go 0 (head points) (tail points
                 else go (bigD + d) point ps newPoints
         go _ _ [] newPoints = reverse newPoints
 
-resampleScala :: Float -> V.Vector (V2 Float) -> V.Vector (V2 Float)
+resampleScala :: Double -> V.Vector (V2 Double) -> V.Vector (V2 Double)
 resampleScala i v@(V.toList -> points) = V.reverse $ V.fromList $ resamplePoints (reverse points) [] 0
     where
-        resamplePoints :: [V2 Float] -> [V2 Float] -> Float -> [V2 Float]
+        resamplePoints :: [V2 Double] -> [V2 Double] -> Double -> [V2 Double]
         resamplePoints [] lz _ = lz
         resamplePoints (h:t) lz d = case t of
             [_] -> lz ++ [h]
@@ -227,11 +227,11 @@ resampleScala i v@(V.toList -> points) = V.reverse $ V.fromList $ resamplePoints
                            in  resamplePoints (q:t) (lz ++ [q]) 0
                       else resamplePoints t lz (d + d')
 
-pathLength :: V.Vector (V2 Float) -> Float
+pathLength :: V.Vector (V2 Double) -> Double
 pathLength v = V.foldl' (\d (prev, next) -> d + distance prev next) 0
              $ V.zipWith (,) v (V.tail v)
 
-test3 :: Float -> [[Float]] -> String
+test3 :: Double -> [[Double]] -> String
 test3 n input = jsifyVector jsifyV2 $ resample n $ readPoints input
 
 
@@ -241,21 +241,21 @@ test3 n input = jsifyVector jsifyV2 $ resample n $ readPoints input
 
 -- determineResampleSpacing
 
-halfwayCornerClean :: V.Vector (V2 Float, Float) -> V2 Float
+halfwayCornerClean :: V.Vector (V2 Double, Double) -> V2 Double
 halfwayCornerClean vec = fst $ V.minimumBy (compare `on` snd) vec'
     where
         vec'    = V.slice quarter len vec
         len     = V.length vec - 2 * quarter
         quarter = V.length vec `div` 4
 
-postProcessCornersClean :: V.Vector (V2 Float, Float) -> V.Vector (V2 Float) -> V.Vector (V2 Float)
+postProcessCornersClean :: V.Vector (V2 Double, Double) -> V.Vector (V2 Double) -> V.Vector (V2 Double)
 postProcessCornersClean points (V.toList -> corners) = V.fromList $ go3 (go1 False corners) []
     where
-        go1 :: Bool -> [V2 Float] -> [V2 Float]
+        go1 :: Bool -> [V2 Double] -> [V2 Double]
         go1 (not -> False) acc = acc
         go1 (not -> True)  acc = go2 True acc []
             where
-                go2 :: Bool -> [V2 Float] -> [V2 Float] -> [V2 Float]
+                go2 :: Bool -> [V2 Double] -> [V2 Double] -> [V2 Double]
                 go2 currentBool corners@(c1:c2:cs) acc =
                     let slice = getSlice points c1 c2
                     in  if not $ isLineClean $ V.map fst slice
@@ -266,27 +266,27 @@ postProcessCornersClean points (V.toList -> corners) = V.fromList $ go3 (go1 Fal
                         else go2 currentBool (c2:cs) (c1:acc)
                 go2 currentBool [c]                acc = go1 currentBool (reverse $ c:acc)
 
-        go3 :: [V2 Float] -> [V2 Float] -> [V2 Float]
+        go3 :: [V2 Double] -> [V2 Double] -> [V2 Double]
         go3 (c1:c:c2:cs) acc = if isLineClean $ V.map fst $ getSlice points c1 c2
                                then go3 (c1:c2:cs) acc
                                else go3 (c:c2:cs) (c1:acc)
         go3 [c1,c2]      acc = reverse $ c2:c1:acc
 
-        getSlice :: V.Vector (V2 Float, Float) -> V2 Float -> V2 Float -> V.Vector (V2 Float, Float)
+        getSlice :: V.Vector (V2 Double, Double) -> V2 Double -> V2 Double -> V.Vector (V2 Double, Double)
         getSlice vec a1 a2 = V.slice i1 (i2 - i1 + 1) vec
             where
                 fstvec = V.map fst vec
                 Just i1 = V.elemIndex a1 fstvec
                 Just i2 = V.elemIndex a2 fstvec
 
-isLineClean :: V.Vector (V2 Float) -> Bool
+isLineClean :: V.Vector (V2 Double) -> Bool
 isLineClean vec = len / pathLen > threshold
     where
         len = distance (V.head vec) (V.last vec)
         pathLen = pathLength vec
         threshold = 0.95
 
-getCornersClean :: V.Vector (V2 Float) -> V.Vector (V2 Float)
+getCornersClean :: V.Vector (V2 Double) -> V.Vector (V2 Double)
 getCornersClean points = runST $ do
     let w = 3
         straws = V.create $ do
@@ -332,7 +332,7 @@ getCornersClean points = runST $ do
 
     return $ postProcessCornersClean (V.zipWith (,) points straws) (V.fromList cornersVal)
 
-shortStrawClean :: V.Vector (V2 Float) -> V.Vector (V2 Float)
+shortStrawClean :: V.Vector (V2 Double) -> V.Vector (V2 Double)
 shortStrawClean points = corners
     where
         s = determineResampleSpacing points
